+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 8.13
 -----
 - Add Liquid Glass support [#4302](https://github.com/Automattic/pocket-casts-ios/pull/4302)
+- Fix an issue with diagonal swipes not dismiss the player
 
 8.12
 -----